feat: Blockscout - styles for tx explanation
