--- src/content/blockscout/components/ExplainBtn/index.module.less
@@ -1,4 +1,0 @@
-.btn {
-  display: flex;
-  align-items: center;
-}
