+++ src/content/blockscout/components/TransactionExplanation/button.module.less
@@ -1,0 +1,24 @@
+.btn {
+  display: flex;
+  align-items: center;
+  margin: 2px 0;
+
+  @media screen and (max-width: 1000px) {
+    margin: 0;
+  }
+
+  font-weight: 600;
+  border-width: 2px;
+  border-radius: 8px;
+  color: var(--chakra-colors-link);
+  border-color: var(--chakra-colors-link);
+  background-color: transparent;
+
+  &:global(.ant-btn-default):global(.ant-btn) {
+    &:hover {
+      color: var(--chakra-colors-link_hovered);
+      border-color: var(--chakra-colors-link_hovered);
+      background-color: transparent;
+    }
+  }
+}
