+++ src/content/blockscout/components/TransactionExplanation/label.module.less
@@ -1,0 +1,16 @@
+.label {
+  display: flex;
+  align-items: center;
+  column-gap: 8px;
+  font-weight: 500;
+  padding: 8px 0;
+  line-height: 20px;
+
+  @media screen and (max-width: 1000px) {
+    padding: 4px 0;
+  }
+}
+
+.logo {
+  margin: 1px !important;
+}
