+++ src/content/blockscout/components/index.ts
@@ -1,4 +1,3 @@
-export { default as ExplainBtn } from './ExplainBtn'
 export { default as MainAddressLabel } from './MainAddressLabel'
 export { default as Tag } from './Tag'
-export { default as TransactionExplanation } from './TransactionExplanation'
+export * as TransactionExplanation from './TransactionExplanation'
