+++ src/content/blockscout/styles/dropdown.less
@@ -1,0 +1,17 @@
+.chakra-ui-light .ant-dropdown-menu {
+  background-color: var(--chakra-colors-white);
+  box-shadow: var(--chakra-shadows-2xl);
+
+  .ant-dropdown-menu-item {
+    color: var(--chakra-colors-black);
+  }
+}
+
+.chakra-ui-dark .ant-dropdown-menu {
+  background-color: var(--chakra-colors-gray-900);
+  box-shadow: var(--chakra-shadows-2xl);
+
+  .ant-dropdown-menu-item {
+    color: var(--chakra-colors-white);
+  }
+}
