+++ src/content/blockscout/styles/index.ts
@@ -1,1 +1,2 @@
 import './tooltip.less'
+import './dropdown.less'
